remove: unused */index.js
