+++ package.json
@@ -68,6 +68,7 @@
     "react": "16.9.0",
     "react-dev-utils": "9.0.4",
     "react-dom": "16.9.0",
+    "typescript": "^4.2.4",
     "webpack": "4.40.2",
     "webpack-node-externals": "1.7.2"
   },
