+++ src/components/Calender/index.js
@@ -1,4 +1,22 @@
-export { default as Calendar } from './Calendar';
-export { default as DateRangePicker } from './DateRangePicker';
-export { default as DatePicker } from './DatePicker';
-export { default as TimePicker } from './TimePicker';
+import React, { forwardRef } from 'react';
+import classNames from 'classnames';
+import CalendarBase from 'react-calendar/dist/entry.nostyle';
+
+function Calendar ({ className, ...props }, ref) {
+  return (
+    <CalendarBase
+      ref={ref}
+      {...props}
+      className={classNames(
+        'Calendar',
+        className && className
+      )}
+    />
+  );
+}
+
+Calendar.displayName = 'Calendar';
+Calendar.defaultProps = {};
+Calendar.propTypes = {};
+
+export default forwardRef(Calendar);
