--- src/constants/index.js
@@ -1,2 +1,0 @@
-export { icons } from './icons';
-export { messagesVariants } from './messages';
